hwpmc(4): Actually use a sufficiently wide type
jhibbits@ points out that left shifting bits 8-11 24 bits won't fit in a 32-bit
integer either.

Corrects r324533.

Submitted by:	jhibbits
Sponsored by:	Dell EMC Isilon


Former-commit-id: b34df781c3220be9b35f4bacbd62261749e1a41a

+++ sys/dev/hwpmc/hwpmc_amd.h
@@ -226,7 +226,7 @@ name|AMD_PMC_TO_EVENTMASK
 parameter_list|(
 name|x
 parameter_list|)
-value|(((x)& 0xFF) | (((uint32_t)(x)& 0xF00)<< 24))
+value|(((x)& 0xFF) | (((uint64_t)(x)& 0xF00)<< 24))
 end_define
 
 begin_define
